This commit adds a Github actions workflow to continuously test that building is successful on Ubuntu 20.04 and Ubuntu 18.04 on every push.
Changes made to CMakeList.txt now support Intel MKL's header file
location installation from apt (on Ubuntu 20.04 and, presumably, later
versions), as well as manual installs using today's current Intel MKL
download and installation through their installer.

However, today's Intel MKL library may not be what you want, and it may
be that the path will change if we use an older version of the library,
but I didn't want to dig that up while I was working on this.