Add io_uring as alternative of aio.
io_uring is enabled by building with -DIOURING.

In the case of aio, io_context_t was defined as the type of IOContext.
io_context_t is a pointer to struct io_context.
io_uring is implemented with the same idea,
defining the type of IOContext as a pointer to struct io_uring.

Signed-off-by: Masahiro Yamada (KIOXIA) <masahiro31.yamada@kioxia.com>